move CloneConfig to its own file
