+++ src/gitingest/schemas/__init__.py
@@ -1,6 +1,7 @@
 """Module containing the schemas for the Gitingest package."""
 
+from gitingest.schemas.cloning import CloneConfig
 from gitingest.schemas.filesystem import FileSystemNode, FileSystemNodeType, FileSystemStats
-from gitingest.schemas.ingestion import CloneConfig, IngestionQuery
+from gitingest.schemas.ingestion import IngestionQuery
 
 __all__ = ["CloneConfig", "FileSystemNode", "FileSystemNodeType", "FileSystemStats", "IngestionQuery"]
